refactor
Took 4 hours 1 minute

+++ xklib-3d/src/main/java/com/xkball/xklib/ui/render/IFont.java
@@ -1,5 +1,6 @@
 package com.xkball.xklib.ui.render;
 
+//todo 宽度为float
 public interface IFont {
     
     int width(String text);
@@ -7,6 +8,6 @@ public interface IFont {
     int lineHeight();
     
     default int width(String text, int lineHeight){
-        return width(text) * (lineHeight / lineHeight());
+        return (int) (width(text) * ((float)lineHeight /(float) lineHeight()));
     }
 }
